Move menu.step.js code to generic.step.js
... because there is no menu.feature anymore
