+++ features/step_definitions/generic.step.js
@@ -199,5 +199,25 @@ defineSupportCode(({ Given, When, Then, setDefaultTimeout }) => {
   When('I scroll to the bottom', () => {
     browser.executeScript('window.scrollBy(0, 10000);');
   });
+
+  When('I click "{itemSelector}" in main menu', (itemSelector, callback) => {
+    waitForElemAndClickIt('.main-menu-icon-button');
+    browser.sleep(1000);
+    waitForElemAndClickIt(`.${itemSelector.replace(/ /g, '-')}`);
+    browser.sleep(1000).then(callback);
+  });
+
+  Then('There is no "{itemSelector}" in main menu', (itemSelector, callback) => {
+    waitForElemAndClickIt('.main-menu-icon-button');
+    browser.sleep(500);
+    expect(element.all(by.css(`md-menu-item .md-button.${itemSelector.replace(/ /g, '-')}`)).count()).to.eventually.equal(0)
+      .and.notify(callback);
+  });
+
+  Then('I should see in "{fieldName}" field:', (fieldName, value, callback) => {
+    const elem = element(by.css(`.${fieldName.replace(/ /g, '-')} textarea`));
+    expect(elem.getAttribute('value')).to.eventually.equal(value)
+      .and.notify(callback);
+  });
 });
